refactor: Context 파일 안에 있는 Provider 함수 따로 분리
