+++ src/App.tsx
@@ -4,18 +4,18 @@ import OperationContainer from '@/components/OperationContainer';
 import useCalculator from '@/hooks/useCalculator';
 
 function App() {
-	const { state } = useCalculator();
+    const { state } = useCalculator();
 
-	return (
-		<div className="App">
-			<div className="calculator">
-				<h1 id="total">{state}</h1>
-				<DigitsContainer/>
-				<ResetContainer/>
-				<OperationContainer/>
-			</div>
-		</div>
-	);
+    return (
+        <div className="App">
+            <div className="calculator">
+                <h1 id="total">{state}</h1>
+                <DigitsContainer/>
+                <ResetContainer/>
+                <OperationContainer/>
+            </div>
+        </div>
+    );
 }
 
 export default App;
