+++ src/hooks/useCalculator.ts
@@ -1,20 +1,20 @@
 import { useContext } from 'react';
-import { CalculatorContext } from '../store/CalculatorContext';
+import { CalculatorContext } from '@/store/CalculatorContext';
 
 export default function useCalculator() {
-	const {
-		insertDigits,
-		insertOperation,
-		setAnswer,
-		resetCalculator,
-		state
-	} = useContext(CalculatorContext);
+    const {
+        insertDigits,
+        insertOperation,
+        setAnswer,
+        resetCalculator,
+        state
+    } = useContext(CalculatorContext);
 
-	return {
-		insertDigits,
-		insertOperation,
-		setAnswer,
-		resetCalculator,
-		state
-	};
+    return {
+        insertDigits,
+        insertOperation,
+        setAnswer,
+        resetCalculator,
+        state
+    };
 }
